+++ Libplanet.Tests/Blocks/BlockTest.cs
@@ -842,5 +842,93 @@ namespace Libplanet.Tests.Blocks
                  () => emptyTxs.Validate(DateTimeOffset.UtcNow)
             );
         }
+
+        [Fact]
+        public void DetectInvalidPreEvaluationHash()
+        {
+            var invalidBlock = new Block<PolymorphicAction<BaseAction>>(
+                index: _fx.Next.Index,
+                difficulty: _fx.Next.Difficulty,
+                totalDifficulty: _fx.Next.TotalDifficulty * 10000, // invalid
+                preEvaluationHash: _fx.Next.PreEvaluationHash,
+                nonce: new Nonce(new byte[] { 0x00 }),
+                miner: _fx.Next.Miner,
+                previousHash: _fx.Next.PreviousHash,
+                timestamp: _fx.Next.Timestamp,
+                transactions: _fx.Next.Transactions);
+
+            Assert.Throws<InvalidBlockPreEvaluationHashException>(() =>
+                invalidBlock.Validate(DateTimeOffset.UtcNow));
+        }
+
+        [Fact]
+        public void DetectInvalidTxSignature()
+        {
+            RawTransaction rawTx = new RawTransaction(
+                0,
+                _fx.TxFixture.Address1.ByteArray,
+                _fx.Genesis.Hash.ByteArray,
+                ImmutableArray<ImmutableArray<byte>>.Empty,
+                _fx.TxFixture.PublicKey1.Format(false).ToImmutableArray(),
+                DateTimeOffset.UtcNow.ToString(
+                    "yyyy-MM-ddTHH:mm:ss.ffffffZ",
+                    CultureInfo.InvariantCulture
+                ),
+                ImmutableArray<IValue>.Empty,
+                new byte[10].ToImmutableArray()
+            );
+            var invalidTx = new Transaction<DumbAction>(rawTx);
+            Assert.Throws<InvalidTxSignatureException>(() =>
+                MineNext(
+                    MineGenesis<DumbAction>(),
+                    new List<Transaction<DumbAction>>
+                    {
+                        invalidTx,
+                    }
+                )
+            );
+        }
+
+        [Fact]
+        public void DetectInvalidTxPublicKey()
+        {
+            RawTransaction rawTxWithoutSig = new RawTransaction(
+                0,
+                new PrivateKey().ToAddress().ByteArray,
+                _fx.Genesis.Hash.ByteArray,
+                ImmutableArray<ImmutableArray<byte>>.Empty,
+                _fx.TxFixture.PublicKey1.Format(false).ToImmutableArray(),
+                DateTimeOffset.UtcNow.ToString(
+                    "yyyy-MM-ddTHH:mm:ss.ffffffZ",
+                    CultureInfo.InvariantCulture
+                ),
+                ImmutableArray<IValue>.Empty,
+                ImmutableArray<byte>.Empty
+            );
+            byte[] sig = _fx.TxFixture.PrivateKey1.Sign(
+                new Transaction<DumbAction>(rawTxWithoutSig).Serialize(false)
+            );
+            var invalidTx = new Transaction<DumbAction>(
+                new RawTransaction(
+                    0,
+                    rawTxWithoutSig.Signer,
+                    rawTxWithoutSig.GenesisHash,
+                    rawTxWithoutSig.UpdatedAddresses,
+                    rawTxWithoutSig.PublicKey,
+                    rawTxWithoutSig.Timestamp,
+                    rawTxWithoutSig.Actions,
+                    sig.ToImmutableArray()
+                )
+            );
+            Assert.Throws<InvalidTxPublicKeyException>(() =>
+                MineNext(
+                    MineGenesis<DumbAction>(),
+                    new List<Transaction<DumbAction>>
+                    {
+                        invalidTx,
+                    }
+                )
+            );
+        }
     }
 }